feat:modeline: Add `chime-enable-modeline` to control display
Introduce a new customizable variable `chime-enable-modeline` to
enable or disable the display of upcoming events in the modeline.
The functionality allows users to set this option to `nil` to
prevent chime from modifying the modeline.

Update relevant functions and tests to respect this new setting.
When `chime-enable-modeline` is `nil`, the modeline string is
cleared or not updated, even when there are upcoming events.